+++ src/Grid.h
@@ -16,7 +16,7 @@ class Grid
         sf::VertexArray m_vLines;
 
 	public:
-		Grid (const sf::Vector2f& gridSize, const sf::Vector2f& pos, const sf::Vector2i& cellSize);
+		Grid (const sf::Vector2f& gridSize, const sf::Vector2f& pos, const sf::Vector2f& cellSize);
 		~Grid();
 
 		void Initialize();
